% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogo_gadget_arms.R
\name{gogo_gadget_arms}
\alias{gogo_gadget_arms}
\title{Gadgetize all the package things}
\usage{
gogo_gadget_arms(package)
}
\arguments{
\item{package}{character vector length 1, the package to gadgetize , must be intalled
will be added to searchPath}
}
\value{
returns nothing but adds A LOT of gadgetized objects to your .GlobalEnv
}
\description{
Gadgetize all the package things
}
\examples{
gogo_gadget_arms("dplyr")

gogo.filter(beaver1, activ == 1)
}
