% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogo_gadget_arms.R
\name{gogo_gadget_arms}
\alias{gogo_gadget_arms}
\title{Gadgetize all the package things}
\usage{
gogo_gadget_arms(package)
}
\arguments{
\item{package}{character vector length 1, the package to gadgetize , must be intalled}
}
\value{
returns nothing but adds an environment to your seach path called gadget_arms
all the exported functions and objects of \code{package} can now be called by typing \code{gogo_{functions/object_name}}
}
\description{
Gadgetize all the package things
}
\examples{

gogo_gadget_arms("dplyr")

gogo_filter(beaver1, activ == 1)

}
